+++ docs/changes.md
@@ -1,6 +1,10 @@
 Changes of joker-cast
 =====================
 
+#### 0.0.19
+* add regular_lookup() and regular_attr_lookup()
+* fix __qualname__ AttributeError in format_function_path()
+
 #### 0.0.18
 * rename number_split() to numseries_segment() with backward compat
 * add nonblank_lines_of()